Split encoding v1 and v2 into two files
