+++ cpp/arcticdb/codec/codec.cpp
@@ -5,7 +5,7 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
-#include <arcticdb/codec/codec_utils.hpp>
+#include <arcticdb/codec/column_encoder_utils.hpp>
 
 
 #include <arcticdb/codec/codec.hpp>
@@ -22,81 +22,13 @@
 #include <google/protobuf/io/zero_copy_stream_impl.h>
 
 
+#include <arcticdb/codec/codec_utils.hpp>
 
 
 namespace arcticdb {
-/// @brief The type to be used for a block which represents shapes
-using ShapesBlock = TypedBlockData<ShapesBlockTDT>;
-
-//TODO  Provide tuning parameters to choose the encoding through encoding tags (also acting as config options)
-template<template<typename> class TypedBlock, class TD, EncodingVersion encoder_version>
-struct TypedBlockEncoderImpl;
-
-template<typename TD, EncodingVersion v>
-using BlockEncoder = TypedBlockEncoderImpl<TypedBlockData, TD, v>;
-
-template<EncodingVersion v>
-struct BytesEncoder {
-    using BytesTypeDescriptorTag = TypeDescriptorTag<DataTypeTag<DataType::UINT8>, DimensionTag<Dimension::Dim1>>;
-    using Encoder = BlockEncoder<BytesTypeDescriptorTag, v>;
-    using BytesBlock = TypedBlockData<BytesTypeDescriptorTag>;
-    using ShapesEncoder = BlockEncoder<ShapesBlockTDT, v>;
-
-    template<typename EncodedFieldType>
-    static void encode(
-        const ChunkedBuffer& data,
-        const arcticdb::proto::encoding::VariantCodec& codec_opts,
-        Buffer& out_buffer,
-        std::ptrdiff_t& pos,
-        EncodedFieldType* encoded_field
-    ) {
-        const shape_t bytes_count = static_cast<shape_t>(data.bytes());
-        if constexpr(v == EncodingVersion::V1) {
-            auto typed_block = BytesBlock(
-                data.data(),
-                &bytes_count,
-                bytes_count,
-                1u,
-                data.block_and_offset(0).block_);
-            Encoder::encode(codec_opts, typed_block, *encoded_field, out_buffer, pos);
-        } else if constexpr(v == EncodingVersion::V2) {
-            const size_t row_count = 1;
-            auto shapes_block = ShapesBlock(&bytes_count,
-                nullptr,
-                sizeof(shape_t),
-                row_count,
-                data.block_and_offset(0).block_);
-            auto data_block = BytesBlock(data.data(),
-                &bytes_count,
-                bytes_count,
-                row_count,
-                data.block_and_offset(0).block_);
-            ShapesEncoder::encode_shapes(shapes_encoding_opts(), shapes_block, *encoded_field, out_buffer, pos);
-            Encoder::encode_values(codec_opts, data_block, *encoded_field, out_buffer, pos);
-            auto* field_nd_array = encoded_field->mutable_ndarray();
-            const auto total_items_count = field_nd_array->items_count() + row_count;
-            field_nd_array->set_items_count(total_items_count);
-        } else {
-            static_assert(std::is_same_v<decltype(v), void>, "Unknown encoding version");
-        }
-    }
-
-    static size_t max_compressed_size(const arcticdb::proto::encoding::VariantCodec& codec_opts, shape_t data_size) {
-        const shape_t shapes_bytes = sizeof(shape_t);
-        const auto values_block = BytesBlock(data_size, &data_size);
-        if constexpr(v == EncodingVersion::V1) {
-            const auto shapes_block = BytesBlock(shapes_bytes, &shapes_bytes);
-            return Encoder::max_compressed_size(codec_opts, values_block) +
-                   Encoder::max_compressed_size(codec_opts, shapes_block);
-        } else if constexpr(v == EncodingVersion::V2) {
-            const auto shapes_block = ShapesBlock(shapes_bytes, &shapes_bytes);
-            return Encoder::max_compressed_size(codec_opts, values_block) +
-                   ShapesEncoder::max_compressed_size(shapes_encoding_opts(), shapes_block);
-        } else {
-            static_assert(std::is_same_v<decltype(v), void>, "Unknown encoding version");
-        }
-    }
-};
+arcticdb::proto::encoding::VariantCodec shapes_encoding_opts() {
+    return codec::default_lz4_codec();
+}
 
 constexpr TypeDescriptor metadata_type_desc() {
     return TypeDescriptor{
@@ -109,350 +41,12 @@ constexpr TypeDescriptor encoded_blocks_type_desc() {
         DataType::UINT8, Dimension::Dim1
     };
 }
+/// @brief The type to be used for a block which represents shapes
+using ShapesBlock = TypedBlockData<ShapesBlockTDT>;
 
-size_t calc_column_blocks_size(const Column& col) {
-    size_t bytes = EncodedField::Size;
-    if(col.type().dimension() != entity::Dimension::Dim0)
-        bytes += sizeof(EncodedBlock);
-
-    bytes += sizeof(EncodedBlock) * col.num_blocks();
-    ARCTICDB_DEBUG(log::version(), "Encoded block size: {} + shapes({}) + {} * {} = {}",
-                 EncodedField::Size,
-                 col.type().dimension() != entity::Dimension::Dim0 ? sizeof(EncodedBlock) : 0u,
-                 sizeof(EncodedBlock),
-                 col.num_blocks(),
-                 bytes);
-
-    return bytes;
-}
-
-size_t encoded_blocks_size(const SegmentInMemory& in_mem_seg) {
-    size_t bytes = 0;
-    for (std::size_t c = 0; c < in_mem_seg.num_columns(); ++c) {
-        const auto& col = in_mem_seg.column(position_t(c));
-        bytes += calc_column_blocks_size(col);
-    }
-    return bytes;
-}
-
-struct SizeResult {
-    size_t max_compressed_bytes_;
-    size_t uncompressed_bytes_;
-    shape_t encoded_blocks_bytes_;
-};
-
-template<EncodingVersion v>
-void calc_metadata_size(
-    const SegmentInMemory& in_mem_seg,
-    const arcticdb::proto::encoding::VariantCodec& codec_opts,
-    SizeResult& result
-    ) {
-    if (in_mem_seg.metadata()) {
-        const auto metadata_bytes = static_cast<shape_t>(in_mem_seg.metadata()->ByteSizeLong());
-        result.uncompressed_bytes_ += metadata_bytes + sizeof(shape_t);
-        result.max_compressed_bytes_ += BytesEncoder<v>::max_compressed_size(codec_opts, metadata_bytes);
-        ARCTICDB_TRACE(log::codec(), "Metadata requires {} max_compressed_bytes", result.max_compressed_bytes_);
-    }
-}
-
-template<EncodingVersion v>
-void calc_columns_size(
-    const SegmentInMemory& in_mem_seg,
-    const arcticdb::proto::encoding::VariantCodec& codec_opts,
-    SizeResult& result
-    ) {
-    ColumnEncoder<v> encoder;
-    for (std::size_t c = 0; c < in_mem_seg.num_columns(); ++c) {
-        auto column_data = in_mem_seg.column_data(c);
-        const auto [uncompressed, required] = encoder.max_compressed_size(codec_opts, column_data);
-        result.uncompressed_bytes_ += uncompressed;
-        result.max_compressed_bytes_ += required;
-        ARCTICDB_TRACE(log::codec(), "Column {} requires {} max_compressed_bytes, total {}", c, required, result.max_compressed_bytes_);
-    }
-}
-
-template<EncodingVersion v>
-void calc_string_pool_size(
-    const SegmentInMemory& in_mem_seg,
-    const arcticdb::proto::encoding::VariantCodec& codec_opts,
-    SizeResult& result
-    ) {
-    if (in_mem_seg.has_string_pool()) {
-        ColumnEncoder<v> encoder;
-        auto string_col = in_mem_seg.string_pool_data();
-        const auto [uncompressed, required] = encoder.max_compressed_size(codec_opts, string_col);
-        result.uncompressed_bytes_ += uncompressed;
-        result.max_compressed_bytes_ += required;
-        ARCTICDB_TRACE(log::codec(), "String pool requires {} max_compressed_bytes, total {}", required, result.max_compressed_bytes_);
-    }
-}
-
-SizeResult max_compressed_size_v1(const SegmentInMemory &in_mem_seg, const arcticdb::proto::encoding::VariantCodec &codec_opts) {
-    ARCTICDB_SAMPLE(GetSegmentCompressedSize, 0)
-    SizeResult result{};
-    calc_metadata_size<EncodingVersion::V1>(in_mem_seg, codec_opts, result);
-
-    if(in_mem_seg.row_count() > 0) {
-        calc_columns_size<EncodingVersion::V1>(in_mem_seg, codec_opts, result);
-        calc_string_pool_size<EncodingVersion::V1>(in_mem_seg, codec_opts, result);
-    }
-    ARCTICDB_TRACE(log::codec(), "Max compressed size {}", result.max_compressed_bytes_);
-    return result;
-}
-
-void calc_encoded_blocks_size(
-    const SegmentInMemory& in_mem_seg,
-    const arcticdb::proto::encoding::VariantCodec& codec_opts,
-    SizeResult& result
-    ) {
-    result.encoded_blocks_bytes_ = static_cast<shape_t>(encoded_blocks_size(in_mem_seg));
-    result.uncompressed_bytes_ += result.encoded_blocks_bytes_;
-    result.max_compressed_bytes_ += BytesEncoder<EncodingVersion::V2>::max_compressed_size(codec_opts, result.encoded_blocks_bytes_);
-}
-
-void calc_stream_descriptor_fields_size(
-    const SegmentInMemory& in_mem_seg,
-    const arcticdb::proto::encoding::VariantCodec& codec_opts,
-    SizeResult& result
-) {
-    auto segment_fields = in_mem_seg.descriptor().fields().column_data();
-    const auto [uncompressed, required] = ColumnEncoderV2::max_compressed_size(codec_opts, segment_fields);
-    result.uncompressed_bytes_ += uncompressed;
-    result.max_compressed_bytes_ += required;
-
-    // Calculate index fields size
-    if(in_mem_seg.index_fields()) {
-        auto index_field_data = in_mem_seg.index_fields()->column_data();
-        const auto [idx_uncompressed, idx_required] = ColumnEncoderV2::max_compressed_size(codec_opts, index_field_data);
-        result.uncompressed_bytes_ += idx_uncompressed;
-        result.max_compressed_bytes_ += idx_required;
-    }
-}
-
-SizeResult max_compressed_size_v2(const SegmentInMemory &in_mem_seg, const arcticdb::proto::encoding::VariantCodec &codec_opts) {
-    ARCTICDB_SAMPLE(GetSegmentCompressedSize, 0)
-    SizeResult result{};
-    result.max_compressed_bytes_ += sizeof(MetadataMagic);
-    calc_metadata_size<EncodingVersion::V2>(in_mem_seg, codec_opts, result);
-    result.max_compressed_bytes_ += sizeof(DescriptorMagic);
-    result.max_compressed_bytes_ += sizeof(IndexMagic);
-    calc_stream_descriptor_fields_size(in_mem_seg, codec_opts, result);
-    result.max_compressed_bytes_ += sizeof(EncodedMagic);
-    calc_encoded_blocks_size(in_mem_seg, codec_opts, result);
-
-    // Calculate fields collection size
-    if(in_mem_seg.row_count() > 0) {
-        result.max_compressed_bytes_ += sizeof(ColumnMagic) * in_mem_seg.descriptor().field_count();
-        calc_columns_size<EncodingVersion::V2>(in_mem_seg, codec_opts, result);
-        result.max_compressed_bytes_ += sizeof(StringPoolMagic);
-        calc_string_pool_size<EncodingVersion::V2>(in_mem_seg, codec_opts, result);
-    }
-    ARCTICDB_TRACE(log::codec(), "Max compressed size {}", result.max_compressed_bytes_);
-    return result;
-}
-
-template<EncodingVersion v>
-void encode_metadata(
-    const SegmentInMemory& in_mem_seg,
-    arcticdb::proto::encoding::SegmentHeader& segment_header,
-    const arcticdb::proto::encoding::VariantCodec& codec_opts,
-    Buffer& out_buffer,
-    std::ptrdiff_t& pos
-    ) {
-    if (in_mem_seg.metadata()) {
-        const auto bytes_count = static_cast<shape_t>(in_mem_seg.metadata()->ByteSizeLong());
-        ARCTICDB_TRACE(log::codec(), "Encoding {} bytes of metadata", bytes_count);
-        auto encoded_field = segment_header.mutable_metadata_field();
-
-        constexpr int max_stack_alloc = 1 << 11;
-        bool malloced{false};
-        uint8_t* meta_ptr{nullptr};
-        if(bytes_count > max_stack_alloc) {
-            meta_ptr = reinterpret_cast<uint8_t*>(malloc(bytes_count));
-            malloced = true;
-        } else {
-            meta_ptr = reinterpret_cast<uint8_t*>(alloca(bytes_count));
-        }
-        ChunkedBuffer meta_buffer;
-        meta_buffer.add_external_block(meta_ptr, bytes_count, 0u);
-        google::protobuf::io::ArrayOutputStream aos(&meta_buffer[0], static_cast<int>(bytes_count));
-        in_mem_seg.metadata()->SerializeToZeroCopyStream(&aos);
-        BytesEncoder<v>::encode(meta_buffer, codec_opts, out_buffer, pos, encoded_field);
-        ARCTICDB_DEBUG(log::codec(), "Encoded metadata to position {}", pos);
-        if(malloced)
-            free(meta_ptr);
-    } else {
-        ARCTICDB_DEBUG(log::codec(), "Not encoding any metadata");
-    }
-}
-
-template<EncodingVersion v>
-void encode_string_pool(
-    const SegmentInMemory& in_mem_seg,
-    arcticdb::proto::encoding::SegmentHeader& segment_header,
-    const arcticdb::proto::encoding::VariantCodec& codec_opts,
-    Buffer& out_buffer,
-    std::ptrdiff_t& pos
-    ) {
-    if (in_mem_seg.has_string_pool()) {
-        ColumnEncoder<v> encoder;
-        ARCTICDB_TRACE(log::codec(), "Encoding string pool to position {}", pos);
-        auto *encoded_field = segment_header.mutable_string_pool_field();
-        auto col = in_mem_seg.string_pool_data();
-        encoder.encode(codec_opts, col, encoded_field, out_buffer, pos);
-        ARCTICDB_TRACE(log::codec(), "Encoded string pool to position {}", pos);
-    }
-}
-
-template<EncodingVersion v, typename = std::enable_if_t<v == EncodingVersion::V2>>
-void encode_field_descriptors(
-    const SegmentInMemory& in_mem_seg,
-    arcticdb::proto::encoding::SegmentHeader& segment_header,
-    const arcticdb::proto::encoding::VariantCodec& codec_opts,
-    Buffer& out_buffer,
-    std::ptrdiff_t& pos
-    ) {
-    ColumnEncoder<v> encoder;
-    ARCTICDB_TRACE(log::codec(), "Encoding field descriptors to position {}", pos);
-    auto *encoded_field = segment_header.mutable_descriptor_field();
-    auto col = in_mem_seg.descriptor().fields().column_data();
-    encoder.encode(codec_opts, col, encoded_field, out_buffer, pos);
-    ARCTICDB_TRACE(log::codec(), "Encoded field descriptors to position {}", pos);
-
-    write_magic<IndexMagic>(out_buffer, pos);
-    if (in_mem_seg.index_fields()) {
-        ARCTICDB_TRACE(log::codec(), "Encoding index fields descriptors to position {}", pos);
-        auto *index_field = segment_header.mutable_index_descriptor_field();
-        auto index_col = in_mem_seg.index_fields()->column_data();
-        encoder.encode(codec_opts, index_col, index_field, out_buffer, pos);
-        ARCTICDB_TRACE(log::codec(), "Encoded index field descriptors to position {}", pos);
-    }
-}
-
-void encode_encoded_fields(
-        arcticdb::proto::encoding::SegmentHeader& segment_header,
-        const arcticdb::proto::encoding::VariantCodec& codec_opts,
-        Buffer& out_buffer,
-        std::ptrdiff_t& pos,
-        const ChunkedBuffer& encoded_blocks_buffer
-        ) {
-
-    ARCTICDB_TRACE(log::codec(), "Encoding encoded blocks to position {}", pos);
-    auto encoded_field = segment_header.mutable_column_fields();
-    encoded_field->set_offset(static_cast<uint32_t>(pos));
-    write_magic<EncodedMagic>(out_buffer, pos);
-	BytesEncoder<EncodingVersion::V2>::encode(encoded_blocks_buffer, codec_opts, out_buffer, pos, encoded_field);
-    ARCTICDB_TRACE(log::codec(), "Encoded encoded blocks to position {}", pos);
-}
-
-Segment encode_v2(SegmentInMemory&& s, const arcticdb::proto::encoding::VariantCodec &codec_opts) {
-    ARCTICDB_SAMPLE(EncodeSegment, 0)
-
-    auto in_mem_seg = std::move(s);
-    auto arena = std::make_unique<google::protobuf::Arena>();
-    auto segment_header = google::protobuf::Arena::CreateMessage<arcticdb::proto::encoding::SegmentHeader>(arena.get());
-    auto& seg_descriptor = in_mem_seg.descriptor();
-    *segment_header->mutable_stream_descriptor() = std::move(seg_descriptor.mutable_proto());
-    segment_header->set_compacted(in_mem_seg.compacted());
-    segment_header->set_encoding_version(static_cast<uint16_t>(EncodingVersion::V2));
-
-    std::ptrdiff_t pos = 0;
-    static auto block_to_header_ratio = ConfigsMap::instance()->get_int("Codec.EstimatedHeaderRatio", 75);
-    const auto preamble = in_mem_seg.num_blocks() * block_to_header_ratio;
-    auto [max_compressed_size, uncompressed_size, encoded_buffer_size] = max_compressed_size_v2(in_mem_seg, codec_opts);
-    ARCTICDB_TRACE(log::codec(), "Estimated max buffer requirement: {}", max_compressed_size);
-    auto out_buffer = std::make_shared<Buffer>(max_compressed_size + encoded_buffer_size, preamble);
-    ARCTICDB_TRACE(log::codec(), "Encoding descriptor: {}", segment_header->stream_descriptor().DebugString());
-    auto *tsd = segment_header->mutable_stream_descriptor();
-    tsd->set_in_bytes(uncompressed_size);
-
-    write_magic<MetadataMagic>(*out_buffer, pos);
-    encode_metadata<EncodingVersion::V2>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
-    write_magic<DescriptorMagic>(*out_buffer, pos);
-    encode_field_descriptors<EncodingVersion::V2>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
-
-    auto encoded_fields_buffer = ChunkedBuffer::presized(static_cast<size_t>(encoded_buffer_size));
-    auto encoded_field_pos = 0u;
-    ColumnEncoder<EncodingVersion::V2> encoder;
-    if(in_mem_seg.row_count() > 0) {
-        ARCTICDB_TRACE(log::codec(), "Encoding fields");
-        for (std::size_t column_index = 0; column_index < in_mem_seg.num_columns(); ++column_index) {
-            write_magic<ColumnMagic>(*out_buffer, pos);
-            auto column_field = new(encoded_fields_buffer.data() + encoded_field_pos) EncodedField;
-            ARCTICDB_TRACE(log::codec(),"Beginning encoding of column {}: ({}) to position {}", column_index, in_mem_seg.descriptor().field(column_index).name(), pos);
-            auto column_data = in_mem_seg.column_data(column_index);
-            encoder.encode(codec_opts, column_data, column_field, *out_buffer, pos);
-            ARCTICDB_TRACE(log::codec(), "Encoded column {}: ({}) to position {}", column_index, in_mem_seg.descriptor().field(column_index).name(), pos);
-            encoded_field_pos += encoded_field_bytes(*column_field);
-            util::check(encoded_field_pos <= encoded_fields_buffer.bytes(),
-                "Encoded field buffer overflow {} > {}",
-                 encoded_field_pos,
-                 encoded_fields_buffer.bytes());
-        }
-        auto field_here ARCTICDB_UNUSED = reinterpret_cast<EncodedField*>(encoded_fields_buffer.data());
-        write_magic<StringPoolMagic>(*out_buffer, pos);
-        encode_string_pool<EncodingVersion::V2>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
-    }
-
-    auto field_before ARCTICDB_UNUSED = reinterpret_cast<EncodedField*>(encoded_fields_buffer.data());
-    encode_encoded_fields(*segment_header, codec_opts, *out_buffer, pos, encoded_fields_buffer);
-    auto field ARCTICDB_UNUSED = reinterpret_cast<EncodedField*>(encoded_fields_buffer.data());
-    out_buffer->set_bytes(pos);
-    tsd->set_out_bytes(pos);
-
-    ARCTICDB_DEBUG(log::codec(), "Encoded header: {}", tsd->DebugString());
-    ARCTICDB_DEBUG(log::codec(), "Block count {} header size {} ratio {}",
-        in_mem_seg.num_blocks(), segment_header->ByteSizeLong(),
-        in_mem_seg.num_blocks() ? segment_header->ByteSizeLong() / in_mem_seg.num_blocks() : 0);
-    return {std::move(arena), segment_header, std::move(out_buffer), seg_descriptor.fields_ptr()};
-}
-
-Segment encode_v1(SegmentInMemory&& s, const arcticdb::proto::encoding::VariantCodec &codec_opts) {
-    /*
-     * This takes an in memory segment with all the metadata, column tensors etc., loops through each column
-     * and based on the type of the column, calls the typed block encoder for that column.
-     */
-    ARCTICDB_SAMPLE(EncodeSegment, 0)
-    auto in_mem_seg = std::move(s);
-    auto arena = std::make_unique<google::protobuf::Arena>();
-    auto segment_header = google::protobuf::Arena::CreateMessage<arcticdb::proto::encoding::SegmentHeader>(arena.get());
-    *segment_header->mutable_stream_descriptor() = in_mem_seg.descriptor().copy_to_proto();
-    segment_header->set_compacted(in_mem_seg.compacted());
-    std::ptrdiff_t pos = 0;
-    static auto block_to_header_ratio = ConfigsMap::instance()->get_int("Codec.EstimatedHeaderRatio", 75);
-    const auto preamble = in_mem_seg.num_blocks() * block_to_header_ratio;
-    auto [max_compressed_size, uncompressed_size, encoded_blocks_bytes] = max_compressed_size_v1(in_mem_seg, codec_opts);
-    ARCTICDB_TRACE(log::codec(), "Estimated max buffer requirement: {}", max_compressed_size);
-    auto out_buffer = std::make_shared<Buffer>(max_compressed_size, preamble);
-    ColumnEncoder<EncodingVersion::V1> encoder;
-
-    ARCTICDB_TRACE(log::codec(), "Encoding descriptor: {}", segment_header->stream_descriptor().DebugString());
-    auto *tsd = segment_header->mutable_stream_descriptor();
-    tsd->set_in_bytes(uncompressed_size);
-
-    encode_metadata<EncodingVersion::V1>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
-
-    if(in_mem_seg.row_count() > 0) {
-        ARCTICDB_TRACE(log::codec(), "Encoding fields");
-        for (std::size_t column_index = 0; column_index < in_mem_seg.num_columns(); ++column_index) {
-            auto column_data = in_mem_seg.column_data(column_index);
-            auto *encoded_field = segment_header->mutable_fields()->Add();
-            encoder.encode(codec_opts, column_data, encoded_field, *out_buffer, pos);
-            ARCTICDB_TRACE(log::codec(), "Encoded column {}: ({}) to position {}", column_index, in_mem_seg.descriptor().fields(column_index).name(), pos);
-        }
-        encode_string_pool<EncodingVersion::V1>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
-    }
-    ARCTICDB_DEBUG(log::codec(), "Setting buffer bytes to {}", pos);
-    out_buffer->set_bytes(pos);
-    tsd->set_out_bytes(pos);
-    ARCTICDB_DEBUG(log::codec(), "Encoded header: {}", tsd->DebugString());
-    if(!segment_header->has_metadata_field())
-        ARCTICDB_DEBUG(log::codec(), "No metadata field");
-    ARCTICDB_DEBUG(log::codec(), "Block count {} header size {} ratio {}",
-                  in_mem_seg.num_blocks(), segment_header->ByteSizeLong(),
-                  in_mem_seg.num_blocks() ? segment_header->ByteSizeLong() / in_mem_seg.num_blocks() : 0);
-    return {std::move(arena), segment_header, std::move(out_buffer), in_mem_seg.descriptor().fields_ptr()};
-}
+//TODO  Provide tuning parameters to choose the encoding through encoding tags (also acting as config options)
+template<template<typename> class TypedBlock, class TD, EncodingVersion encoder_version>
+struct TypedBlockEncoderImpl;
 
 namespace {
 class MetaBuffer {
