+++ cpp/arcticdb/codec/codec_v1.cpp
@@ -4,15 +4,26 @@
  *
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
-
-#include <arcticdb/codec/codec_v1.hpp>
-#include <arcticdb/codec/encoded_field.hpp>
 #include <arcticdb/codec/codec_utils.hpp>
 #include <arcticdb/codec/typed_block_encoder_impl.hpp>
-#include <arcticdb/column_store/column_data.hpp>
+#include <arcticdb/codec/column_encoder_utils.hpp>
+#include <arcticdb/column_store/memory_segment.hpp>
 
-#include <utility>
 namespace arcticdb {
+    /// @brief Utility class used to encode and compute the max encoding size for regular data columns for V1 encoding
+    struct ColumnEncoderV1 {
+        static std::pair<size_t, size_t> max_compressed_size(
+            const arcticdb::proto::encoding::VariantCodec& codec_opts,
+            ColumnData& column_data);
+
+        static void encode(
+            const arcticdb::proto::encoding::VariantCodec &codec_opts,
+            ColumnData& column_data,
+            std::variant<EncodedField*, arcticdb::proto::encoding::EncodedField*> variant_field,
+            Buffer& out,
+            std::ptrdiff_t& pos);
+    };
+
     std::pair<size_t, size_t> ColumnEncoderV1::max_compressed_size(
         const arcticdb::proto::encoding::VariantCodec& codec_opts,
         ColumnData& column_data
@@ -60,5 +71,65 @@ namespace arcticdb {
             }
         });
         encode_sparse_map(column_data, variant_field, out, pos);
+    }
+
+    SizeResult max_compressed_size_v1(const SegmentInMemory &in_mem_seg, const arcticdb::proto::encoding::VariantCodec &codec_opts) {
+        ARCTICDB_SAMPLE(GetSegmentCompressedSize, 0)
+        SizeResult result{};
+        calc_metadata_size<EncodingVersion::V1>(in_mem_seg, codec_opts, result);
+
+        if(in_mem_seg.row_count() > 0) {
+            calc_columns_size<ColumnEncoderV1>(in_mem_seg, codec_opts, result);
+            calc_string_pool_size<ColumnEncoderV1>(in_mem_seg, codec_opts, result);
+        }
+        ARCTICDB_TRACE(log::codec(), "Max compressed size {}", result.max_compressed_bytes_);
+        return result;
+    }
+
+    Segment encode_v1(SegmentInMemory&& s, const arcticdb::proto::encoding::VariantCodec &codec_opts) {
+        /*
+         * This takes an in memory segment with all the metadata, column tensors etc., loops through each column
+         * and based on the type of the column, calls the typed block encoder for that column.
+         */
+        ARCTICDB_SAMPLE(EncodeSegment, 0)
+        auto in_mem_seg = std::move(s);
+        auto arena = std::make_unique<google::protobuf::Arena>();
+        auto segment_header = google::protobuf::Arena::CreateMessage<arcticdb::proto::encoding::SegmentHeader>(arena.get());
+        *segment_header->mutable_stream_descriptor() = in_mem_seg.descriptor().copy_to_proto();
+        segment_header->set_compacted(in_mem_seg.compacted());
+        std::ptrdiff_t pos = 0;
+        static auto block_to_header_ratio = ConfigsMap::instance()->get_int("Codec.EstimatedHeaderRatio", 75);
+        const auto preamble = in_mem_seg.num_blocks() * block_to_header_ratio;
+        auto [max_compressed_size, uncompressed_size, encoded_blocks_bytes] = max_compressed_size_v1(in_mem_seg, codec_opts);
+        ARCTICDB_TRACE(log::codec(), "Estimated max buffer requirement: {}", max_compressed_size);
+        auto out_buffer = std::make_shared<Buffer>(max_compressed_size, preamble);
+        ColumnEncoderV1 encoder;
+
+        ARCTICDB_TRACE(log::codec(), "Encoding descriptor: {}", segment_header->stream_descriptor().DebugString());
+        auto *tsd = segment_header->mutable_stream_descriptor();
+        tsd->set_in_bytes(uncompressed_size);
+
+        encode_metadata<EncodingVersion::V1>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
+
+        if(in_mem_seg.row_count() > 0) {
+            ARCTICDB_TRACE(log::codec(), "Encoding fields");
+            for (std::size_t column_index = 0; column_index < in_mem_seg.num_columns(); ++column_index) {
+                auto column_data = in_mem_seg.column_data(column_index);
+                auto *encoded_field = segment_header->mutable_fields()->Add();
+                encoder.encode(codec_opts, column_data, encoded_field, *out_buffer, pos);
+                ARCTICDB_TRACE(log::codec(), "Encoded column {}: ({}) to position {}", column_index, in_mem_seg.descriptor().fields(column_index).name(), pos);
+            }
+            encode_string_pool<ColumnEncoderV1>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
+        }
+        ARCTICDB_DEBUG(log::codec(), "Setting buffer bytes to {}", pos);
+        out_buffer->set_bytes(pos);
+        tsd->set_out_bytes(pos);
+        ARCTICDB_DEBUG(log::codec(), "Encoded header: {}", tsd->DebugString());
+        if(!segment_header->has_metadata_field())
+            ARCTICDB_DEBUG(log::codec(), "No metadata field");
+        ARCTICDB_DEBUG(log::codec(), "Block count {} header size {} ratio {}",
+            in_mem_seg.num_blocks(), segment_header->ByteSizeLong(),
+            in_mem_seg.num_blocks() ? segment_header->ByteSizeLong() / in_mem_seg.num_blocks() : 0);
+        return {std::move(arena), segment_header, std::move(out_buffer), in_mem_seg.descriptor().fields_ptr()};
     }
 }
