+++ cpp/arcticdb/codec/codec_v2.cpp
@@ -5,19 +5,59 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
-#include <arcticdb/codec/codec_v2.hpp>
 #include <arcticdb/codec/encoded_field.hpp>
 #include <arcticdb/codec/codec_utils.hpp>
 #include <arcticdb/codec/typed_block_encoder_impl.hpp>
-#include <arcticdb/column_store/column_data.hpp>
+#include <arcticdb/codec/column_encoder_utils.hpp>
+#include <arcticdb/column_store/memory_segment.hpp>
 
-#include <utility>
 
 namespace arcticdb {
-    using ShapesBlockTDT = TypeDescriptorTag<DataTypeTag<DataType::INT64>, DimensionTag<Dimension::Dim0>>;
-    using ShapesBlock = TypedBlockData<ShapesBlockTDT>;
+    /// TODO: Move
+    using DescriptorMagic = util::MagicNum<'D','e','s','c'>;
+    using EncodedMagic = util::MagicNum<'E','n','c','d'>;
+    using StringPoolMagic = util::MagicNum<'S','t','r','p'>;
+    using MetadataMagic = util::MagicNum<'M','e','t','a'>;
+    using IndexMagic = util::MagicNum<'I','n','d','x'>;
+    using ColumnMagic = util::MagicNum<'C','l','m','n'>;
 
-    static ShapesBlock create_shapes_typed_block(const ColumnData& column_data) {
+    using ShapesBlockTDT = TypeDescriptorTag<DataTypeTag<DataType::INT64>, DimensionTag<Dimension::Dim0>>;
+
+    /// @brief Utility class used to encode and compute the max encoding size for regular data columns for V2 encoding
+    /// What differs this from the already existing ColumnEncoder is that ColumnEncoder encodes the shapes of
+    /// multidimensional data as part of each block. ColumnEncoder2 uses a better strategy and encodes the shapes for the
+    /// whole column upfront (before all blocks).
+    /// @note Although ArcticDB did not support multidimensional user data prior to creating ColumnEncoder2 some of the
+    /// internal data was multidimensional and used ColumnEncoder. More specifically: string pool and metadata.
+    /// @note This should be used for V2 encoding. V1 encoding can't use it as there is already data written the other
+    ///	way and it will be hard to distinguish both.
+    struct ColumnEncoderV2 {
+    public:
+        static void encode(
+            const arcticdb::proto::encoding::VariantCodec &codec_opts,
+            ColumnData& column_data,
+            std::variant<EncodedField*, arcticdb::proto::encoding::EncodedField*> variant_field,
+            Buffer& out,
+            std::ptrdiff_t& pos);
+        static std::pair<size_t, size_t> max_compressed_size(
+            const arcticdb::proto::encoding::VariantCodec& codec_opts,
+            ColumnData& column_data);
+    private:
+        static void encode_shapes(
+            const ColumnData& column_data,
+            std::variant<EncodedField*, arcticdb::proto::encoding::EncodedField*> variant_field,
+            Buffer& out,
+            std::ptrdiff_t& pos_in_buffer);
+        static void encode_blocks(
+            const arcticdb::proto::encoding::VariantCodec &codec_opts,
+            ColumnData& column_data,
+            std::variant<EncodedField*, arcticdb::proto::encoding::EncodedField*> variant_field,
+            Buffer& out,
+            std::ptrdiff_t& pos);
+    };
+
+
+    static TypedBlockData<ShapesBlockTDT> create_shapes_typed_block(const ColumnData& column_data) {
         static_assert(std::is_same_v<ShapesBlockTDT::DataTypeTag::raw_type, shape_t>,
             "Shape block type is not compatible");
         const size_t row_count = column_data.shapes()->bytes() / sizeof(shape_t);
@@ -50,7 +90,7 @@ namespace arcticdb {
         // assign 0 for the shape upon reading. There is one edge case - when we have None in the column, as it should not
         // have shape at all (since it's not an array). This is handled by the sparse map.
         if(column_data.type().dimension() != Dimension::Dim0 && !is_empty_type(column_data.type().data_type())) {
-            ShapesBlock shapes_block = create_shapes_typed_block(column_data);
+            TypedBlockData<ShapesBlockTDT> shapes_block = create_shapes_typed_block(column_data);
             util::variant_match(variant_field, [&](auto field){
                 using ShapesEncoder = TypedBlockEncoderImpl<TypedBlockData, ShapesBlockTDT, EncodingVersion::V2>;
                 ShapesEncoder::encode_shapes(shapes_encoding_opts(), shapes_block, *field, out, pos_in_buffer);
@@ -92,7 +132,7 @@ namespace arcticdb {
             using ShapesEncoder = TypedBlockEncoderImpl<TypedBlockData, ShapesBlockTDT, EncodingVersion::V2>;
             ARCTICDB_TRACE(log::codec(), "Column data has {} blocks", column_data.num_blocks());
             const size_t shapes_byte_count = column_data.shapes()->bytes();
-            const ShapesBlock shapes_block = create_shapes_typed_block(column_data);
+            const TypedBlockData<ShapesBlockTDT> shapes_block = create_shapes_typed_block(column_data);
             max_compressed_bytes += ShapesEncoder::max_compressed_size(shapes_encoding_opts(), shapes_block);
             uncompressed_bytes += shapes_byte_count;
             while (auto block = column_data.next<TDT>()) {
@@ -109,5 +149,188 @@ namespace arcticdb {
             add_bitmagic_compressed_size(column_data, uncompressed_bytes, max_compressed_bytes);
             return std::make_pair(uncompressed_bytes, max_compressed_bytes);
         });
+    }
+
+    template<typename MagicType>
+    void write_magic(Buffer& buffer, std::ptrdiff_t& pos) {
+        new (buffer.data() + pos) MagicType{};
+        pos += sizeof(MagicType);
+    }
+
+    void encode_field_descriptors(
+        const SegmentInMemory& in_mem_seg,
+        arcticdb::proto::encoding::SegmentHeader& segment_header,
+        const arcticdb::proto::encoding::VariantCodec& codec_opts,
+        Buffer& out_buffer,
+        std::ptrdiff_t& pos
+    ) {
+        ARCTICDB_TRACE(log::codec(), "Encoding field descriptors to position {}", pos);
+        auto *encoded_field = segment_header.mutable_descriptor_field();
+        auto col = in_mem_seg.descriptor().fields().column_data();
+        ColumnEncoderV2::encode(codec_opts, col, encoded_field, out_buffer, pos);
+        ARCTICDB_TRACE(log::codec(), "Encoded field descriptors to position {}", pos);
+
+        write_magic<IndexMagic>(out_buffer, pos);
+        if (in_mem_seg.index_fields()) {
+            ARCTICDB_TRACE(log::codec(), "Encoding index fields descriptors to position {}", pos);
+            auto *index_field = segment_header.mutable_index_descriptor_field();
+            auto index_col = in_mem_seg.index_fields()->column_data();
+            ColumnEncoderV2::encode(codec_opts, index_col, index_field, out_buffer, pos);
+            ARCTICDB_TRACE(log::codec(), "Encoded index field descriptors to position {}", pos);
+        }
+    }
+
+    size_t calc_column_blocks_size(const Column& col) {
+        size_t bytes = EncodedField::Size;
+        if(col.type().dimension() != entity::Dimension::Dim0)
+            bytes += sizeof(EncodedBlock);
+
+        bytes += sizeof(EncodedBlock) * col.num_blocks();
+        ARCTICDB_DEBUG(log::version(), "Encoded block size: {} + shapes({}) + {} * {} = {}",
+            EncodedField::Size,
+            col.type().dimension() != entity::Dimension::Dim0 ? sizeof(EncodedBlock) : 0u,
+            sizeof(EncodedBlock),
+            col.num_blocks(),
+            bytes);
+
+        return bytes;
+    }
+
+    size_t encoded_blocks_size(const SegmentInMemory& in_mem_seg) {
+        size_t bytes = 0;
+        for (std::size_t c = 0; c < in_mem_seg.num_columns(); ++c) {
+            const auto& col = in_mem_seg.column(position_t(c));
+            bytes += calc_column_blocks_size(col);
+        }
+        return bytes;
+    }
+
+    void calc_encoded_blocks_size(
+        const SegmentInMemory& in_mem_seg,
+        const arcticdb::proto::encoding::VariantCodec& codec_opts,
+        SizeResult& result
+    ) {
+        result.encoded_blocks_bytes_ = static_cast<shape_t>(encoded_blocks_size(in_mem_seg));
+        result.uncompressed_bytes_ += result.encoded_blocks_bytes_;
+        result.max_compressed_bytes_ += BytesEncoder<EncodingVersion::V2>::max_compressed_size(codec_opts, result.encoded_blocks_bytes_);
+    }
+
+    void calc_stream_descriptor_fields_size(
+        const SegmentInMemory& in_mem_seg,
+        const arcticdb::proto::encoding::VariantCodec& codec_opts,
+        SizeResult& result
+    ) {
+        auto segment_fields = in_mem_seg.descriptor().fields().column_data();
+        const auto [uncompressed, required] = ColumnEncoderV2::max_compressed_size(codec_opts, segment_fields);
+        result.uncompressed_bytes_ += uncompressed;
+        result.max_compressed_bytes_ += required;
+
+        // Calculate index fields size
+        if(in_mem_seg.index_fields()) {
+            auto index_field_data = in_mem_seg.index_fields()->column_data();
+            const auto [idx_uncompressed, idx_required] = ColumnEncoderV2::max_compressed_size(codec_opts, index_field_data);
+            result.uncompressed_bytes_ += idx_uncompressed;
+            result.max_compressed_bytes_ += idx_required;
+        }
+    }
+
+    SizeResult max_compressed_size_v2(const SegmentInMemory &in_mem_seg, const arcticdb::proto::encoding::VariantCodec &codec_opts) {
+        ARCTICDB_SAMPLE(GetSegmentCompressedSize, 0)
+        SizeResult result{};
+        result.max_compressed_bytes_ += sizeof(MetadataMagic);
+        calc_metadata_size<EncodingVersion::V2>(in_mem_seg, codec_opts, result);
+        result.max_compressed_bytes_ += sizeof(DescriptorMagic);
+        result.max_compressed_bytes_ += sizeof(IndexMagic);
+        calc_stream_descriptor_fields_size(in_mem_seg, codec_opts, result);
+        result.max_compressed_bytes_ += sizeof(EncodedMagic);
+        calc_encoded_blocks_size(in_mem_seg, codec_opts, result);
+
+        // Calculate fields collection size
+        if(in_mem_seg.row_count() > 0) {
+            result.max_compressed_bytes_ += sizeof(ColumnMagic) * in_mem_seg.descriptor().field_count();
+            calc_columns_size<ColumnEncoderV2>(in_mem_seg, codec_opts, result);
+            result.max_compressed_bytes_ += sizeof(StringPoolMagic);
+            calc_string_pool_size<ColumnEncoderV2>(in_mem_seg, codec_opts, result);
+        }
+        ARCTICDB_TRACE(log::codec(), "Max compressed size {}", result.max_compressed_bytes_);
+        return result;
+    }
+
+    void encode_encoded_fields(
+        arcticdb::proto::encoding::SegmentHeader& segment_header,
+        const arcticdb::proto::encoding::VariantCodec& codec_opts,
+        Buffer& out_buffer,
+        std::ptrdiff_t& pos,
+        const ChunkedBuffer& encoded_blocks_buffer
+    ) {
+
+        ARCTICDB_TRACE(log::codec(), "Encoding encoded blocks to position {}", pos);
+        auto encoded_field = segment_header.mutable_column_fields();
+        encoded_field->set_offset(static_cast<uint32_t>(pos));
+        write_magic<EncodedMagic>(out_buffer, pos);
+        BytesEncoder<EncodingVersion::V2>::encode(encoded_blocks_buffer, codec_opts, out_buffer, pos, encoded_field);
+        ARCTICDB_TRACE(log::codec(), "Encoded encoded blocks to position {}", pos);
+    }
+
+    Segment encode_v2(SegmentInMemory&& s, const arcticdb::proto::encoding::VariantCodec &codec_opts) {
+        ARCTICDB_SAMPLE(EncodeSegment, 0)
+
+        auto in_mem_seg = std::move(s);
+        auto arena = std::make_unique<google::protobuf::Arena>();
+        auto segment_header = google::protobuf::Arena::CreateMessage<arcticdb::proto::encoding::SegmentHeader>(arena.get());
+        auto& seg_descriptor = in_mem_seg.descriptor();
+        *segment_header->mutable_stream_descriptor() = std::move(seg_descriptor.mutable_proto());
+        segment_header->set_compacted(in_mem_seg.compacted());
+        segment_header->set_encoding_version(static_cast<uint16_t>(EncodingVersion::V2));
+
+        std::ptrdiff_t pos = 0;
+        static auto block_to_header_ratio = ConfigsMap::instance()->get_int("Codec.EstimatedHeaderRatio", 75);
+        const auto preamble = in_mem_seg.num_blocks() * block_to_header_ratio;
+        auto [max_compressed_size, uncompressed_size, encoded_buffer_size] = max_compressed_size_v2(in_mem_seg, codec_opts);
+        ARCTICDB_TRACE(log::codec(), "Estimated max buffer requirement: {}", max_compressed_size);
+        auto out_buffer = std::make_shared<Buffer>(max_compressed_size + encoded_buffer_size, preamble);
+        ARCTICDB_TRACE(log::codec(), "Encoding descriptor: {}", segment_header->stream_descriptor().DebugString());
+        auto *tsd = segment_header->mutable_stream_descriptor();
+        tsd->set_in_bytes(uncompressed_size);
+
+        write_magic<MetadataMagic>(*out_buffer, pos);
+        encode_metadata<EncodingVersion::V2>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
+        write_magic<DescriptorMagic>(*out_buffer, pos);
+        encode_field_descriptors(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
+
+        auto encoded_fields_buffer = ChunkedBuffer::presized(static_cast<size_t>(encoded_buffer_size));
+        auto encoded_field_pos = 0u;
+        ColumnEncoderV2 encoder;
+        if(in_mem_seg.row_count() > 0) {
+            ARCTICDB_TRACE(log::codec(), "Encoding fields");
+            for (std::size_t column_index = 0; column_index < in_mem_seg.num_columns(); ++column_index) {
+                write_magic<ColumnMagic>(*out_buffer, pos);
+                auto column_field = new(encoded_fields_buffer.data() + encoded_field_pos) EncodedField;
+                ARCTICDB_TRACE(log::codec(),"Beginning encoding of column {}: ({}) to position {}", column_index, in_mem_seg.descriptor().field(column_index).name(), pos);
+                auto column_data = in_mem_seg.column_data(column_index);
+                encoder.encode(codec_opts, column_data, column_field, *out_buffer, pos);
+                ARCTICDB_TRACE(log::codec(), "Encoded column {}: ({}) to position {}", column_index, in_mem_seg.descriptor().field(column_index).name(), pos);
+                encoded_field_pos += encoded_field_bytes(*column_field);
+                util::check(encoded_field_pos <= encoded_fields_buffer.bytes(),
+                    "Encoded field buffer overflow {} > {}",
+                    encoded_field_pos,
+                    encoded_fields_buffer.bytes());
+            }
+            auto field_here ARCTICDB_UNUSED = reinterpret_cast<EncodedField*>(encoded_fields_buffer.data());
+            write_magic<StringPoolMagic>(*out_buffer, pos);
+            encode_string_pool<ColumnEncoderV2>(in_mem_seg, *segment_header, codec_opts, *out_buffer, pos);
+        }
+
+        auto field_before ARCTICDB_UNUSED = reinterpret_cast<EncodedField*>(encoded_fields_buffer.data());
+        encode_encoded_fields(*segment_header, codec_opts, *out_buffer, pos, encoded_fields_buffer);
+        auto field ARCTICDB_UNUSED = reinterpret_cast<EncodedField*>(encoded_fields_buffer.data());
+        out_buffer->set_bytes(pos);
+        tsd->set_out_bytes(pos);
+
+        ARCTICDB_DEBUG(log::codec(), "Encoded header: {}", tsd->DebugString());
+        ARCTICDB_DEBUG(log::codec(), "Block count {} header size {} ratio {}",
+            in_mem_seg.num_blocks(), segment_header->ByteSizeLong(),
+            in_mem_seg.num_blocks() ? segment_header->ByteSizeLong() / in_mem_seg.num_blocks() : 0);
+        return {std::move(arena), segment_header, std::move(out_buffer), seg_descriptor.fields_ptr()};
     }
 }
